backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
Pillow>=10.3.0
websockets>=12.0
google-generativeai==0.7.2
